+++ packages/actions-iso/tsconfig.json
@@ -1,0 +1,60 @@
+{
+  "compilerOptions": {
+    "outDir": "build",
+    "target": "es5",
+    "module": "esnext",
+    "moduleResolution": "node",
+    "jsx": "react",
+    // optimised
+    "removeComments": true,
+    // non-strict rules
+    "forceConsistentCasingInFileNames": true,
+    "noFallthroughCasesInSwitch": false,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "noImplicitAny": false,
+    "allowUnreachableCode": true,
+    "alwaysStrict": false,
+    "strict": false,
+    // strict rules
+    "noImplicitReturns": false,
+    "noImplicitThis": true,
+    // other
+    "isolatedModules": true,
+    // "allowJs": true,
+    "checkJs": false,
+    "skipLibCheck": true,
+    "sourceMap": true,
+    "resolveJsonModule": true,
+    // good pratice
+    "allowSyntheticDefaultImports": true,
+    "declaration": true,
+    "declarationMap": true,
+    "esModuleInterop": true,
+    "importHelpers": true,
+    "experimentalDecorators": true,
+    "baseUrl": ".",
+
+    // "paths": {
+    //   "@jd/common/*": ["packages/common/src/*"],
+    // },
+    "lib": ["dom", "dom.iterable", "esnext"]
+  },
+  "typeRoots": ["node_modules/@types"],
+
+  "exclude": [
+    "node_modules",
+    "**/node_modules",
+    "build",
+    "**/build",
+    "dist/**",
+    "**/dist",
+    "scripts",
+    "jest",
+    "__tests__",
+    "**/__tests__",
+    "__jest__",
+    "**/__jest__",
+    "config/*"
+  ]
+}
